Add reselect, extract navigation logic into helper
