+++ .babelrc
@@ -8,10 +8,12 @@
                 "@components": "./src/components",
                 "@config": "./src/config",
                 "@constants": "./src/constants",
+                "@helpers": "./src/helpers",
                 "@img": "./src/img",
                 "@plugins": "./src/plugins",
                 "@reducers": "./src/reducers",
                 "@schema": "./src/schema",
+                "@selectors": "./src/selectors",
                 "@services": "./src/services",
                 "@validators": "./src/validators",
                 "@views": "./src/views"
