+++ .eslintrc.yml
@@ -39,10 +39,12 @@ settings:
         '@components': './src/components'
         '@config': './src/config'
         '@constants': './src/constants'
+        '@helpers': './src/helpers'
         '@img': './src/img'
         '@plugins': './src/plugins'
         '@reducers': './src/reducers'
         '@schema': './src/schema'
+        '@selectors': './src/selectors'
         '@services': './src/services'
         '@validators': './src/validators'
         '@views': './src/views'
