+++ package.json
@@ -42,6 +42,7 @@
     "redux-mock-store": "^1.3.0",
     "redux-thunk": "^2.1.0",
     "remove-markdown": "^0.3.0",
+    "reselect": "^4.0.0",
     "smoothscroll-polyfill": "^0.3.6",
     "uswds": "^1.6.9"
   },
