+++ src/selectors/formType.js
@@ -1,0 +1,4 @@
+import { createSelector } from 'reselect'
+
+export const formTypeSelector = state => state.authentication.formType
+
